+++ sdk/trade360-java-sdk/src/test/java/eu/lsports/trade360_java_sdk/common/entities/outright_league/OutrightLeagueFixtureTest.java
@@ -19,6 +19,7 @@ class OutrightLeagueFixtureTest {
         assertNull(fixture.subscription);
         assertNull(fixture.sport);
         assertNull(fixture.location);
+        assertNull(fixture.startDate);
         assertNull(fixture.lastUpdate);
         assertNull(fixture.endDate);
         assertNull(fixture.status);
@@ -27,7 +28,9 @@ class OutrightLeagueFixtureTest {
         fixture.subscription = new Subscription();
         fixture.sport = new Sport();
         fixture.location = new Location();
+        Instant start = Instant.now().minus(1, ChronoUnit.HOURS);
         Instant now = Instant.now();
+        fixture.startDate = start;
         fixture.lastUpdate = now;
         fixture.status = FixtureStatus.FINISHED;
         fixture.endDate = Instant.now().plus(2, ChronoUnit.HOURS);
@@ -40,6 +43,7 @@ class OutrightLeagueFixtureTest {
         assertNotNull(fixture.subscription);
         assertNotNull(fixture.sport);
         assertNotNull(fixture.location);
+        assertEquals(start, fixture.startDate);
         assertEquals(now, fixture.lastUpdate);
         assertEquals(FixtureStatus.FINISHED, fixture.status);
         assertNotNull(fixture.extraData);